Safari-611.1.10/LayoutTests/fast/css/getComputedStyle/counterIncrement-without-counter.html

<html>
<head>
<script src="../../../resources/js-test-pre.js"></script>
</head>
<body>
<div id="console"></div>
<script>
var style = window.getComputedStyle(document.body, null);
shouldBeEqualToString("style.counterIncrement", "none");
shouldBeEqualToString("style.counterReset", "none");
</script>
</body>
</html>
