Safari-611.1.10/LayoutTests/fast/css/getComputedStyle/counterIncrement-without-counter-expected.txt

PASS style.counterIncrement is "none"
PASS style.counterReset is "none"
